+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/product/productDetail/ProductDetailUiState.kt
@@ -1,6 +1,9 @@
 package com.app.arcabyolimpo.presentation.screens.product.productDetail
 
+import com.app.arcabyolimpo.domain.model.product.Product
+
 data class ProductDetailUiState(
+    val product: Product? = null,
     val isLoading: Boolean = false,
     val error: String? = null,
     val decisionDialogVisible: Boolean = false,